Add `Distribution` metric type
This adds a new `Distribution` metric type and specialized support for it in the reporter (distributions are sent to a different API endpoint than all other metrics). I packed the specialized support into the reporter so this had a minimal impact on the rest of the architecture (everything still flushes by just returning an array of objects formatted for the API).