Use ranges for dependency versions
Both dependencies of this package used to exact dependencies, which can result in a lot of redundancy if a consumer of this package has other dependencies that also use `debug` or `@datadog/datadog-api-client`. Instead, we now use "compatible with" ranges to let consumers use any version of our dependencies that should work with our code.

This also updates `debug` to the current major release. (This declares anything compatible with 4.1, not 4.0, since 4.1 has an important fix.)