Fix the broken `apiHost` option
It looks like the `apiHost` option was added in #50, had some fixes committed directly to the main branch later (?), and then was broken in #62 when the underlying Datadog client implementation was changed. This fixes it, and resolves #42.

The `agent` option was also broken in #62, but the PR discussion makes that look intentional (while there are no notes about breaking `apiHost`). However, only part of the implementation awas removed and it was still documented in some spots. I also cleaned all that out here.