Update test tools (#94)
This updates Mocha and Chai (and plugins) to their latest compatible versions and fixes #87.

Mocha v10 (the latest) drops support for Node.js v12, so we can't update to it. Instead, we can update to the latest v9.x release, which at least fixes a number of security vulnerabilities. Our usage of Mocha isn't too complex so happily none of our tests need to change for this.

On the other hand, there are some big breaking changes in Chai v4, so this involves a bunch of changes to the tests. The most significant thing is that `deep` now always and only means to do a deep equality check on the value being compared to, while a new `nested` property always means that the provided *keys* might be keys or might be paths. In v3.x, `deep` was used for both in different circumstances, which led to some fuzziness about what a given assertion is doing.