Update Chai and plugins to current major version
There are some big breaking changes in Chai v4, so this involves a bunch of changes to the tests. The most significant thing is that `deep` now always and only means to do a deep equality check on the value being compared to, while a new `nested` property always means that the provided *keys* might be keys or might be paths. In v3.x, `deep` was used for both in different circumstances, which led to some fuzziness about what a given assertion is doing.