Never silence sending errors
If metrics failed while sending to Datadog and no explicit error handler was set, log the error regardless of the user's DEBUG setting. This is usually critical info and should not ever be silenced. Fixes #52.

This also creates a central place to configure `debug` so we don't ever have typos or other configuration inconsistencies across the package.