Build TypeScript types from doc comments
This converts a number of comments in the code to JSDoc-style comments (and adds some more detailed descriptions/types along the way) and uses TypeScript to build type definition files from the comments. This allows you to publish types directly from this package instead of making TypeScript user install `@types/datadog-metrics`, which is maintained separately, by hand, and won't always be up-to-date.

It does not add new comments for objects that are not publicly exposed (although that would be a good idea for consistency). It also adds NPM lifecycle scripts to automatically build types when publishing.

Fixes #86.