Rename `apiHost` → `site` and `DataDogReporter` → `DatadogReporter` (#100)
The `apiHost` property here is the same as what Datadog’s docs refer to `site`, so we’ve renamed it to match. We’ve also lower-cased the second "d" in `DatadogReporter`, since the previous spelling was wrong (see #59 and #98 for more about that).

Both of the old names still work, but using them prints deprecation warnings to let users know they should update their code.