Export a *type* for BufferedMetricsLogger
Even though we started generating TypeScript types from the JSDoc a while back, it turns out `BufferedMetricsLogger` was getting exported as a value and not a type, which technically works in TS, but is definitely not what you want. This changes up our syntax a bit to get the compiler to do the right thing, and adds a basic test of the types to make sure we don't regress.

Fixes #119.