Export a type for BufferedMetricsLogger (#120)
Even though we started generating TypeScript types from the JSDoc a while back, it turns out `BufferedMetricsLogger` was getting exported as a *value* and not a type. That technically works in TS, but requires some weird syntax and is definitely not what you want. This changes up our imports and exports a bit to get the TypeScript compiler to do the right thing, and adds a basic test of the types to make sure we don't regress.

See #119.