Add live API integration check script
This basic script sends some metrics to Datadog using this library, then queries the API for those same values to make sure they were successfully ingested and will be visible in Datadog's UI. This is mainly meant as a simple test to ensure changes to the underlying Datadog client keep sending things in the correct form. Unit tests *mostly* cover this use case, so the check here is pretty lightweight. We don't check every metric type, for example. But this is still important.

Fixes #89.