Rename `DATADOG_API_HOST` → `DATADOG_SITE`
Back in #100, I renamed the `apiHost` option to `site` (in order to match up with Datadog's own docs and terminology), but forgot to rename the corresponding environment variable. This deprecates `DATADOG_API_HOST` in favor of `DATADOG_SITE` (as should have been done before). We'll remove support for the old env var when we remove `apiHost`.