Add script to test live API integration (#128)
This new `npm run check-integration` script sends some metrics and distributions to Datadog using this library, then queries the API for those same metric values to make sure they were successfully ingested and will be visible in Datadog's UI.

The goal here is to provide a minimal check of real, live integration in case a change to the actual reporter or Datadog client breaks or mis-formats something in a way that our unit tests don’t effectively catch. In *most* cases, our unit tests should be enough (this is why we use nock instead of just mocking the Datadog client), so this check is pretty lightweight (it only checks one metric type, for example).

This also moves some other test tools as scripts from the `test` to the `test-other` directory so they don’t get mixed up with unit tests.