The simplest possible retry implementation
This relies on the built-in retry functionality added to `@datadog/datadog-api-client` to retry API requests on HTTP errors. It keeps things simple, but the main drawback is that it does not retry on *network* errors (when the Datadog server on the other end is unable to receive or reply to the request in the first place). There's an open issue for that on the Datadog client repo, but no replies.

Fixes #108.