First pass at auto-flushing on process exit
This is an ugly first-pass at auto-flushing before the process exits. This previously required manually flushing, which is easy to forget if everything just happens automatically otherwise. Fixes #136.

There's a lot that should be cleaned up here (the `BufferedMetricsLogger` just has way too much logic now, for example). But this gets it working and adds tests. This also adds a `close()` method (maybe it should be renamed? stop? shutdown?) that disables autoflushing and flushes any buffered metrics.