Auto-flush on process exit (#141)
This attaches to the process’s `beforeExit` event (supported in Node.js, Deno, and Bun) when auto-flushing is enabled in order to automatically flush when a program exits. Before, you needed to remember to call `metrics.flush()` manually at the end of your program, which is easy to forget (or not even know about in the first place!).

It also adds a `stop()` method that disables auto-flushing and flushes any remaining buffered metrics.

Fixes #136.