Drop the official Datadog client (#144)
This drops our dependency on `@datadog/datadog-api-client` in favor of directly calling the API via HTTP. It requires a little more code on our side, but gets us a bit more flexibility and drops a 15 MB (!) dependency that we barely use.

This partly covers #132. There is definitely followup work to do here, but we want to get this out the door so people can use it and provide feedback.